FEAT - Função de deletar um post de usuário logado.
