+++ package.json
@@ -11,6 +11,7 @@
         "react-debounce-input": "^3.3.0",
         "react-dom": "^17.0.2",
         "react-icons": "^4.4.0",
+        "react-modal": "^3.15.1",
         "react-router-dom": "^6.3.0",
         "react-scripts": "^4.0.3",
         "styled-components": "^5.3.5",
