+++ src/components/Feed/NewPost/index.jsx
@@ -1,8 +1,8 @@
 import { useContext, useState } from 'react';
-import DataContext from '../../hooks/DataContext';
+import DataContext from '../../../hooks/DataContext';
 
 import { Wrapper, PostForm } from './styles';
-import Axios from '../../blueprints';
+import Axios from '../../../blueprints';
 
 const mockAvatar = 'https://avatars.githubusercontent.com/u/90518458?v=4';
