+++ src/components/Feed/NewPost/styles.js
